Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -69,7 +69,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Acceso denegado");    
+        }        
         return view("questions.edit", compact('question'));
+        
     }
 
     /**
@@ -81,6 +85,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Acceso denegado");    
+        }   
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', "Your question has been updated.");
@@ -94,6 +101,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, "Acceso denegado");    
+        }   
         $question->delete();
 
         return redirect('/questions/')->with('success', "Your question has been deleted");
